requirements.txt
django
djangorestframework
pillow
python-dotenv
mysqlclient
InstagramApi
